requirements/base.txt
https://www.djangoproject.com/download/1.7b1/tarball/
http://github.com/django-compressor/django-compressor/tarball/develop/
django-tastypie==0.11.0
django-filemanager
